+++ custom_components/ecoflow_energy/coordinator/core.py
@@ -41,6 +41,7 @@ from ..const import (
     DEVICE_TYPE_DISPLAY_NAMES,
     DEVICE_TYPE_SMARTPLUG,
     DEVICE_TYPE_STREAM,
+    DEVICE_TYPE_STREAM_5000,
     DEVICE_TYPE_STREAM_AC5000,
     DEVICE_TYPE_UNKNOWN,
     DOMAIN,
@@ -312,7 +313,7 @@ class EcoFlowDeviceCoordinator(
         elif self.device_type == DEVICE_TYPE_STREAM:
             self._power_to_energy = STREAM_POWER_TO_ENERGY
             self._energy_from_api = STREAM_ENERGY_FROM_API
-        elif self.device_type == DEVICE_TYPE_STREAM_AC5000:
+        elif self.device_type in (DEVICE_TYPE_STREAM_AC5000, DEVICE_TYPE_STREAM_5000):
             self._power_to_energy = STREAMAC5000_POWER_TO_ENERGY
             self._energy_from_api = STREAMAC5000_ENERGY_FROM_API
         else: